requirements.txt
django==3.2
psycopg2==2.8.5
django-environ==0.8.1
djangorestframework==3.12.4
pytz==2022.1
drf-yasg==1.20.0
pillow==9.0.1
factory-boy==3.2.1

django-extensions==3.1.5
ipython==8.2.0
black==22.1.0